requirements.txt
Django==4.2.4
djangorestframework==3.14.0
psycopg2-binary==2.9.7

pytest~=7.4.0
django-filter~=23.2